# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Logistic API."
  version: "1.0.0"
  title: "LOL Foods API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "rafael.sjp17@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "logistic-app"
basePath: "/api"
schemes:
  - "https"
  - "http"
paths:
  /restaurant:
    post:
      tags:
        - "Restaurants"
      summary: "create a restaurant"
      description: ""
      operationId: "addRestaurant"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Restaurant object that needs to be added to the database"
          required: true
          schema:
            $ref: "#/definitions/Restaurant"
      responses:
        "201":
          description: "Created"
        "400":
          description: "Bad request"
        "401":
          description: "Unauthorized"
        "405":
          description: "Invalid input"
        "500":
          description: "Internal server error"
      security:
        - petstore_auth:
            - "write:pets"
            - "read:pets"

    put:
      tags:
        - "Restaurants"
      summary: "Update an existing restaurant"
      description: ""
      operationId: "updateRestaurant"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Restaurant object to be updated"
          required: true
          schema:
            $ref: "#/definitions/Restaurant"
      responses:
        "200":
          description: "OK"
        "400":
          description: "Bad request"
        "401":
          description: "Unauthorized"
        "405":
          description: "Invalid input"
        "500":
          description: "Internal server error"

      security:
        - petstore_auth:
            - "write:pets"
            - "read:pets"
  /restaurant/{id}:
    get:
      tags:
        - "Restaurants"
      summary: "Finds Restaurant by id"
      operationId: "findRestaurantById"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Status values that need to be considered for filter"
          required: true
          type: "integer"

      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Restaurant"
        "204":
          description: "No content"
        "400":
          description: "Bad request"
        "401":
          description: "Unauthorized"
        "405":
          description: "Invalid input"
        "500":
          description: "Internal server error"

      security:
        - petstore_auth:
            - "write:pets"
            - "read:pets"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Restaurant:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 5
      lat:
        type: number
        format: double
        example: 25.698
      lon:
        type: number
        format: double
        example: -8.785
...